# dashboard/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Flaglane API",
    "description" : "Feature flags and remote configuration. `/api/**` serves the dashboard with a JWT; `/sdk/**` serves SDKs with an API key.",
    "license" : {
      "name" : "Apache License 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "0.1.0"
  },
...
  "paths" : {
...
    "/api/projects/{projectKey}/flags" : {
      "get" : {
        "tags" : [ "Flags" ],
        "summary" : "List a project's flags",
        "description" : "Archived flags included.",
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The project's flags",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FlagResponse"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "No such project or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Flags" ],
        "summary" : "Create a flag",
        "description" : "Every environment gets a configuration: disabled, falling through to `false`, at 0% (FR-FLG-003).",
        "operationId" : "create_1",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateFlagRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FlagResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The key is used by a live or an archived flag",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
    "/api/projects/{projectKey}/flags/{flagKey}/restore" : {
      "post" : {
        "tags" : [ "Flags" ],
        "summary" : "Restore an archived flag",
        "description" : "Same key, same configurations (FR-FLG-007).",
        "operationId" : "restore",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The flag, restored",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FlagResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
    "/api/projects/{projectKey}/flags/{flagKey}/archive" : {
      "post" : {
        "tags" : [ "Flags" ],
        "summary" : "Archive a flag",
        "description" : "Removes it from every ruleset. Never a deletion: its history stays, its key stays reserved, and it can be restored (FR-FLG-005).",
        "operationId" : "archive",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The flag, archived",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FlagResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
    "/api/projects/{projectKey}/flags/{flagKey}" : {
      "get" : {
        "tags" : [ "Flags" ],
        "summary" : "Get a flag",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The flag",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FlagResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "Flags" ],
        "summary" : "Edit a flag",
        "description" : "Name, description and client-side visibility. The key never changes.",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateFlagRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "409" : {
            "description" : "The flag is archived",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "The flag as edited",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FlagResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ProblemDetail" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri"
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : [ "string", "null" ]
          },
          "instance" : {
            "type" : [ "string", "null" ],
            "format" : "uri"
          },
          "properties" : {
            "type" : [ "object", "null" ],
            "additionalProperties" : { }
          }
        }
      },
...
      "CreateFlagRequest" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string",
            "description" : "Unique in the project, archived flags included; can never change",
            "example" : "new-checkout",
            "pattern" : "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
          },
          "name" : {
            "type" : "string",
            "example" : "New checkout",
            "maxLength" : 100,
            "minLength" : 0
          },
          "description" : {
            "type" : [ "string", "null" ],
            "maxLength" : 1000,
            "minLength" : 0
          },
          "clientSideVisible" : {
            "type" : "boolean",
            "default" : false,
            "description" : "Whether client keys may read it. Defaults to false: browser keys are public"
          }
        },
        "required" : [ "key", "name" ]
      },
      "FlagResponse" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string",
            "example" : "new-checkout"
          },
          "name" : {
            "type" : "string",
            "example" : "New checkout"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "clientSideVisible" : {
            "type" : "boolean"
          },
          "archivedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time",
            "description" : "Set while the flag is archived"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "archivedAt", "clientSideVisible", "createdAt", "description", "key", "name" ]
      },
...
      "UpdateFlagRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : [ "string", "null" ],
            "maxLength" : 100,
            "minLength" : 1
          },
          "description" : {
            "type" : [ "string", "null" ],
            "description" : "Empty clears it",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "clientSideVisible" : {
            "type" : [ "boolean", "null" ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "description" : "An access token from `POST /api/auth/login`, valid for 8 hours.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "sdk-key" : {
        "type" : "http",
        "description" : "An API key issued for one environment. A server key reads every flag; a client key reads client-side-visible flags only, with no user overrides.",
        "scheme" : "bearer",
        "bearerFormat" : "flg_srv_… or flg_cli_…"
      }
    }
  }
}